orga deploy.sh

+++ deploy.sh
@@ -1,6 +1,10 @@
 #!/usr/bin/env bash
 
-if sudo lsof -ti :8081; then
+port_in_use() {
+    sudo lsof -ti :$1
+}
+
+if port_in_use 8081; then
     IDLE_PROFILE=set2
 else
     IDLE_PROFILE=set1
@@ -29,6 +33,8 @@ echo ">>> $JAR_NAME 를 profile=$IDLE_PROFILE 로 실행."
 
 nohup java -jar -Dspring.profiles.active=$IDLE_PROFILE $JAR_NAME > $REPOSITORY/nohup.out 2>&1 &
 
+sleep 10
+
 echo ">>> 전환한 port로 nginx 설정 변경"
 echo "set \$service_url http://127.0.0.1:${IDLE_PORT};" | sudo tee /etc/nginx/conf.d/service-url.inc
 echo ">>> restart Nginx"
@@ -43,9 +49,11 @@ fi
 echo ">>> 프로세스 종료할 이전 버전의 port: $KILL_TO_PORT"
 echo ">>> $KILL_TO_PORT PID 확인"
 
-KILL_PID=$(sudo lsof -ti tcp:${KILL_TO_PORT})
-
-echo ">>> kill -15 $KILL_PID"
-kill -15 ${KILL_PID}
-sleep 5
-
+if port_in_use $KILL_TO_PORT; then
+    KILL_PID=$(sudo lsof -ti tcp:${KILL_TO_PORT})
+    echo ">>> kill -15 $KILL_PID"
+    kill -15 ${KILL_PID}
+    sleep 5
+else
+    echo ">>> 이전 버전의 프로세스가 없습니다."
+fi
